requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
pydantic
passlib[bcrypt]
python-jose
email-validator
fastapi-mail
cloudinary
slowapi
python-multipart
